move tsx out of main
